[#5] Move ApiCredential to model directory
