+++ data/src/main/java/co/nimblehq/chorn/survey/data/model/ApiCredential.kt
@@ -1,4 +1,4 @@
-package co.nimblehq.chorn.survey.data.service
+package co.nimblehq.chorn.survey.data.model
 
 data class ApiCredential(
     val clientId: String,
